+++ tcs/src/main/java/go/party/tcs/repository/NotificationRepository.java
@@ -12,6 +12,6 @@ public interface NotificationRepository extends JpaRepository<Notification, Long
     
     List<Notification> findByUserId(Integer userId);
 
-    long countByUserIdAndVisualizadoFalse(Integer userId);
+    int countByUserIdAndVisualizadoFalse(Integer userId);
 
 }
